Fix autoboxing warning
git-svn-id: https://svn.apache.org/repos/asf/tomcat/trunk@1492167 13f79535-47bb-0310-9956-ffa450edef68

+++ java/org/apache/catalina/tribes/group/interceptors/TcpFailureDetector.java
@@ -276,7 +276,7 @@ public class TcpFailureDetector extends ChannelInterceptorBase {
             } else {
                 if (removeSuspectsTimeout > 0) {
                     long timeNow = System.currentTimeMillis();
-                    int timeIdle = (int) ((timeNow - removeSuspects.get(m)) / 1000L);
+                    int timeIdle = (int) ((timeNow - removeSuspects.get(m).longValue()) / 1000L);
                     if (timeIdle > removeSuspectsTimeout) {
                         removeSuspects.remove(m); // remove suspect member
                     }
